refactor variables that setup.py requires to separate files
This modifies the codebase that we don't (currently) need anything
outside of the stdlib to run setup.py; this makes it possible to `pip
install` the repo without first installing some of its dependencies
manually.
